third_party/blink/web_tests/external/wpt/css/css-images/parsing/object-fit-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['object-fit'] = "contain scale-down" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['object-fit'] = "cover scale-down" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['object-fit'] = "scale-down contain" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['object-fit'] = "scale-down cover" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
